feat : EmailVerified error solve
