+++ tcono/build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'java'
 }
 
-group = 'blog.econovation'
+group = 'com.econovation'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '11'
 
@@ -22,6 +22,10 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-data-jdbc'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
+
+//	update null field setting
+	implementation 'org.mapstruct:mapstruct:1.4.1.Final'
+	annotationProcessor 'org.mapstruct:mapstruct-processor:1.4.1.Final'
 
 //	SpringBoot Mail Setting
 	implementation 'org.springframework.boot:spring-boot-starter-mail'
