--- tcono/src/main/java/blog/econovation/tcono/TconoApplication.java
@@ -1,13 +1,0 @@
-package blog.econovation.tcono;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing // JPA Auditing 활성화
-@SpringBootApplication
-public class TconoApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(TconoApplication.class, args);
-	}
-}
