--- tcono/src/main/java/blog/econovation/tcono/domain/auth/ConfirmationTokenRepository.java
@@ -1,13 +1,0 @@
-package blog.econovation.tcono.domain.auth;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.time.LocalDateTime;
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Long> {
-    Optional<ConfirmationToken> findByIdAndExpirationDateAfterAndExpired(UUID confirmationTokenId, LocalDateTime now, boolean expired);
-}
