--- tcono/src/main/java/blog/econovation/tcono/domain/user/Role.java
@@ -1,11 +1,0 @@
-package blog.econovation.tcono.domain.user;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-    ADMIN,GUEST,USER;
-}
