--- tcono/src/main/java/blog/econovation/tcono/service/BadRequestException.java
@@ -1,11 +1,0 @@
-package blog.econovation.tcono.service;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(code = HttpStatus.BAD_REQUEST, reason = "error.bad")
-public class BadRequestException extends RuntimeException {
-    public BadRequestException(String token_not_found) {
-
-    }
-}
