--- tcono/src/main/java/blog/econovation/tcono/web/controller/SessionConst.java
@@ -1,5 +1,0 @@
-package blog.econovation.tcono.web.controller;
-
-public class SessionConst {
-    public static final String LOGIN_MEMBER = "loginMember";
-}
