--- tcono/src/main/java/blog/econovation/tcono/web/dto/UserResponseDto.java
@@ -1,33 +1,0 @@
-package blog.econovation.tcono.web.dto;
-
-import blog.econovation.tcono.domain.user.User;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Data
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class UserResponseDto {
-    private Long year;
-    private String userName;
-    private String password;
-    private String userEmail;
-
-    @Builder
-    public UserResponseDto(Long year, String userName, String password, String userEmail) {
-        this.year = year;
-        this.userName = userName;
-        this.password = password;
-        this.userEmail = userEmail;
-    }
-
-    public User toEntity(){
-        return User.builder()
-                .userEmail(userEmail)
-                .year(year)
-                .userName(userEmail).build();
-    }
-}
