+++ tcono/src/main/java/com/econovation/tcono/WebConfig.java
@@ -1,0 +1,26 @@
+package com.econovation.tcono;
+
+import com.econovation.tcono.Interceptor.LogInterceptor;
+import com.econovation.tcono.Interceptor.LoginCheckInterceptor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new LogInterceptor())
+                .order(1)
+                .addPathPatterns("/**")
+                .excludePathPatterns("/css/**", "/*.ico", "/error", "/api/user/**","/api/confirm-email/**");
+
+        registry.addInterceptor(new LoginCheckInterceptor())
+                .order(2)
+                .addPathPatterns("/**")
+                .excludePathPatterns("/", "/api/user/**","/api/user","/api/confirm-email/**", "/api/login", "/api/logout",
+                        "/css/**", "/*.ico", "/error");
+    }
+}
