+++ tcono/src/main/java/com/econovation/tcono/domain/auth/ConfirmationTokenRepository.java
@@ -1,0 +1,13 @@
+package com.econovation.tcono.domain.auth;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Long> {
+    Optional<ConfirmationToken> findByIdAndExpirationDateAfterAndExpired(UUID confirmationTokenId, LocalDateTime now, boolean expired);
+}
