+++ tcono/src/main/java/com/econovation/tcono/domain/user/Role.java
@@ -1,0 +1,11 @@
+package com.econovation.tcono.domain.user;
+
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+    ADMIN,GUEST,USER;
+}
