+++ tcono/src/main/java/com/econovation/tcono/web/SessionConst.java
@@ -1,0 +1,5 @@
+package com.econovation.tcono.web;
+
+public class SessionConst {
+    public static final String LOGIN_MEMBER = "loginMember";
+}
