+++ tcono/src/main/java/com/econovation/tcono/web/controller/SessionConst.java
@@ -1,0 +1,5 @@
+package com.econovation.tcono.web.controller;
+
+public class SessionConst {
+    public static final String LOGIN_MEMBER = "loginMember";
+}
