+++ tcono/src/main/java/com/econovation/tcono/web/dto/UserUpdateRequestDto.java
@@ -1,0 +1,36 @@
+package com.econovation.tcono.web.dto;
+
+import com.econovation.tcono.domain.user.User;
+import lombok.Data;
+import lombok.Getter;
+import org.hibernate.validator.constraints.Range;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
+@Data
+@Getter
+public class UserUpdateRequestDto {
+    @NotBlank
+    private String userEmail;
+
+    @NotNull
+    @Range(min =1, max = 50)
+    private Long year;
+
+    @NotNull
+    private String userName;
+
+    public UserUpdateRequestDto(String userEmail, Long year, String userName) {
+        this.userEmail = userEmail;
+        this.year = year;
+        this.userName = userName;
+    }
+
+    public User toEntity(){
+        return User.builder()
+                .userEmail(userEmail)
+                .year(year)
+                .userName(userName).build();
+    }
+}
