+++ tcono/src/main/resources/application.properties
@@ -12,7 +12,7 @@ graphql.tools.schema-location-pattern=**/*.graphqls
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
 spring.mail.username=ymecca730135@gmail.com
-spring.mail.password=usezwvpmppuokxsw
+spring.mail.password=bpvdqvlitrzeqpfb
 spring.mail.properties.mail.smtp.starttls.enable=true
 spring.mail.properties.mail.smtp.starttls.required=true
 spring.mail.properties.mail.smtp.auth=true
@@ -23,7 +23,7 @@ spring.datasource.driver-class-name=org.h2.Driver
 spring.jpa.generate-ddl=true
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
 
 spring.mvc.hiddenmethod.filter.enabled=true
