--- tcono/src/test/java/blog/econovation/tcono/TconoApplicationTests.java
@@ -1,13 +1,0 @@
-package blog.econovation.tcono;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TconoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
