+++ tcono/src/test/java/com/econovation/tcono/TconoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.econovation.tcono;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TconoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
